Struct transient_hashmap::TransientHashMap

pub struct TransientHashMap<K, V, T = StandardTimer> where
    T: Timer,  { /* fields omitted */ }

HashMap with entries that will be garbage collected (pruned) after not being used for specified time.

Pruning does not occur automatically, make sure to call prune method to remove old entries.

Methods

impl<K, V> TransientHashMap<K, V, StandardTimer> where
K: Eq + Hash + Clone,

pub fn new(lifetime: u32) -> Self

Creates new TransientHashMap with standard timer and specified entries lifetime.

impl<K, V, T> TransientHashMap<K, V, T> where
K: Eq + Hash + Clone,
T: Timer,

pub fn new_with_timer(lifetime: u32, t: T) -> Self

Creates new TransientHashMap with given timer and specfied entries lifetime.

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Insert new entry to this map overwriting any previous entry.

Prolongs lifetime of key.

pub fn entry(&mut self, key: K) -> Entry<K, V>

Insert new entry to this map overwriting any previous entry.

Always prolongs the lifetime of key. TODO [ToDr] Should only prolong if new item is inserted or entry is occupied.

pub fn get(&mut self, key: &K) -> Option<&V>

Gets reference to stored value.

Prolongs lifetime of key if is in the map.

pub fn get_mut(&mut self, key: &K) -> Option<&mut V>

Gets mutable reference to stored value.

Prolongs lifetime of key if is in the map.

pub fn contains_key(&mut self, key: &K) -> bool

Checks if key is contained.

Prolongs lifetime of key if is in the map.

pub fn remove(&mut self, k: &K) -> Option<V>

Removes key from the map if present.

Also removes associated lifetime.

pub fn remaining_lifetime(&mut self, key: &K) -> Option<u32>

Returns remaining lifetime of key without altering it.

pub fn prune(&mut self) -> Vec<K>

Clear overdue entries from the TransientHashMap.

pub fn direct(&self) -> &HashMap<K, V>

Get a reference to backing HashMap.

pub fn direct_mut(&mut self) -> &mut HashMap<K, V>

Get the mutable reference to backing HashMap.